fix: rename update switch to polling switch
The `update switch` was confusing since we use `update available` to describe whether a software update is available. We now clarify that this switch controls polling of a vehicle.

BREAKING CHANGE: `update_switch` has been renamed to `polling_switch`. While the UI name will change immediately if you have not modified it, the entity_id should not change unless you remove and reinstall the component.

closes #22

Co-authored-by: Andy Allsopp <arallsopp@gmail.com>