Disable forced updated for device trackers
Forced updates are only needed if we are not polling. Since the coordinator
is effectively doing the polling instead of Home Assistant internals doing it
via should_poll set to True, we need to set the property manually to avoid
writing a state update every time the coordinator callback happens to avoid
a state changed event when nothing has really changed.