fix: change timestamp calculation
While charging the timestamp would continuously update. Now the value
will only update if charging, the remaining time is greater than 0 and
the timestamp is more than a minute different. The timestamp will report
unavailable if charging_state is unknown or stopped. It will remember
the last value only if charging completed.