feat: Sync car state from Teslamate MQTT
Since Teslamate doesn't map directly to TeslaCar from teslajsonpy,
we need to have special handling of both 'plugged_in' and 'state'
when they come from Teslamate. When we receive a new 'plugged_in'-value,
we check the `car.state` to determine the new 'charging_state'
in teslajsonpy, and similarly we need to check the `car.charging_state`
when Teslamate sends a new car-state.

Also refactor the MQTT-code slightly to avoid multiple implementations
of the code that updates the TeslaCar-object.