Only modify charging state when car state is 'charging'
The previous code would always update the charging state when
receiving a new car state. This could sometimes cause the
charging state to be set to "Disconnected" even when the car was
physically connected. Avoid this by only changing the charging
state if the new received car state is "charging".
Otherwise don't modify the charging state of the car. Hopefully this
should lead to consistent behavior.