Avoid writing state if poll did not happen
#948 made me realize that the car data was being polled far less
frequently than I thought as the underlying library was caching, but
every time the coordinator fired, it would still callback all the listeners
and write the state of all the entities which meant we ended up writing
state every 10 seconds even if nothing has changed.

Since the data the coordinator returns can be compared with __eq__, enabling
always_update=False will check if the data has changed and only fire
the listeners if it has. This avoids 1000s of entity writes
per minute which was the bulk of the state writes on my HA instance